+++ packages/design-system-react-native/src/components/ButtonBase/ButtonBase.stories.tsx
@@ -1,9 +1,11 @@
-import { BoxFlexDirection } from '@metamask/design-system-shared';
+import {
+  BoxFlexDirection,
+  ButtonBaseSize,
+  IconName,
+} from '@metamask/design-system-shared';
 import type { Meta, StoryObj } from '@storybook/react-native';
 
-import { ButtonBaseSize } from '../../types';
 import { Box } from '../Box';
-import { IconName } from '../Icon';
 
 import { ButtonBase } from './ButtonBase';
 import type { ButtonBaseProps } from './ButtonBase.types';