+++ packages/design-system-react-native/src/components/ButtonIcon/ButtonIcon.test.tsx
@@ -1,10 +1,12 @@
+import {
+  ButtonIconSize,
+  IconName,
+  ButtonIconVariant,
+} from '@metamask/design-system-shared';
 import { useTailwind } from '@metamask/design-system-twrnc-preset';
 import { renderHook } from '@testing-library/react-hooks';
 import { render, fireEvent } from '@testing-library/react-native';
 import React from 'react';
-
-import { ButtonIconSize, ButtonIconVariant } from '../../types';
-import { IconName } from '../Icon';
 
 import { ButtonIcon } from './ButtonIcon';
 import { TWCLASSMAP_BUTTONICON_SIZE_DIMENSION } from './ButtonIcon.constants';