+++ packages/design-system-shared/src/index.ts
@@ -77,6 +77,17 @@ export {
 // ButtonFilter types (ADR-0004)
 export { type ButtonFilterPropsShared } from './types/ButtonFilter';
 
+// ButtonBase types (ADR-0003 + ADR-0004)
+export {
+  ButtonBaseSize,
+  ButtonHeroSize,
+  ButtonSize,
+  type ButtonBasePropsShared,
+} from './types/ButtonBase';
+
+// Button types (ADR-0003 + ADR-0004)
+export { ButtonVariant, type ButtonPropsShared } from './types/Button';
+
 // AvatarNetwork types (ADR-0004)
 export { type AvatarNetworkPropsShared } from './types/AvatarNetwork';
 
@@ -105,6 +116,12 @@ export {
   type BadgeWrapperCustomPosition,
   type BadgeWrapperPropsShared,
 } from './types/BadgeWrapper';
+// ButtonIcon types (ADR-0003 + ADR-0004)
+export {
+  ButtonIconSize,
+  ButtonIconVariant,
+  type ButtonIconPropsShared,
+} from './types/ButtonIcon';
 
 export {
   AvatarAccountSize,
@@ -121,6 +138,9 @@ export {
   TextVariant,
   type TextPropsShared,
 } from './types/Text';
+
+// TextField types (ADR-0004)
+export { type TextFieldPropsShared } from './types/TextField';
 
 // AvatarFavicon types (ADR-0004)
 export {