+++ packages/design-system-shared/src/types/ButtonIcon/index.ts
@@ -1,0 +1,5 @@
+export {
+  ButtonIconSize,
+  ButtonIconVariant,
+  type ButtonIconPropsShared,
+} from './ButtonIcon.types';