+++ packages/design-system-shared/src/types/Icon/Icon.types.ts
@@ -294,6 +294,7 @@ export const IconName = {
   TableRow: 'TableRow',
   Tablet: 'Tablet',
   Tag: 'Tag',
+  Telegram: 'Telegram',
   ThumbDownFilled: 'ThumbDownFilled',
   ThumbDown: 'ThumbDown',
   ThumbUpFilled: 'ThumbUpFilled',